+++ src/commands/general/help.js
@@ -1,0 +1,6 @@
+export default {
+  name: 'help',
+  aliases: ['h'],
+  usage: '**!help**',
+  description: 'Help',
+};
