+++ src/commands/lol/api-version.js
@@ -1,0 +1,12 @@
+import { getApiVersion } from '../../utils/lol';
+
+export default {
+  name: 'api-version',
+  aliases: ['api'],
+  usage: '**!lol api-version**',
+  description: 'Returns the api version',
+  execute: async ({ msg }) => {
+    const apiVersion = await getApiVersion();
+    msg.channel.send(`I'm currently using api version ${apiVersion}`);
+  },
+};
